Abstracting SpecialTileLay: Creation of a new Class SpecialSingleTileLay after abstracting the methods from SpecialTileLay.configureFromXML Fixing the Configurations of the implemented Variants to Call SpecicalSingleTileLay instead of SpecialTileLay.
