+++ src/main/java/net/sf/rails/game/special/SpecialTileLay.java
@@ -11,17 +11,17 @@ import net.sf.rails.game.*;
 import net.sf.rails.util.Util;
 
 
-public class SpecialTileLay extends SpecialProperty {
+public abstract class SpecialTileLay extends SpecialProperty {
 
-    private String locationCodes = null;
-    private List<MapHex> locations = null;
-    private String tileId = null;
-    private Tile tile = null;
-    private String name = null;
-    private boolean extra = false;
-    private boolean free = false;
-    private int discount = 0;
-    private boolean connected = false;
+    protected String locationCodes = null;
+    protected List<MapHex> locations = null;
+    protected String tileId = null;
+    protected Tile tile = null;
+    protected String name = null;
+    protected boolean extra = false;
+    protected boolean free = false;
+    protected int discount = 0;
+    protected boolean connected = false;
     
     /** Tile colours that can be laid with this special property.
      * Default is same colours as is allowed in a a normal tile lay.
@@ -38,76 +38,13 @@ public class SpecialTileLay extends SpecialProperty {
     @Override
     public void configureFromXML(Tag tag) throws ConfigurationException {
         super.configureFromXML(tag);
-
-        Tag tileLayTag = tag.getChild("SpecialTileLay");
-        if (tileLayTag == null) {
-            throw new ConfigurationException("<SpecialTileLay> tag missing");
-        }
-
-        locationCodes = tileLayTag.getAttributeAsString("location");
-        if (!Util.hasValue(locationCodes))
-            throw new ConfigurationException("SpecialTileLay: location missing");
-
-        tileId = tileLayTag.getAttributeAsString("tile", null);
-
-        String coloursString = tileLayTag.getAttributeAsString("colour");
-        if (Util.hasValue(coloursString)) {
-            tileColours = coloursString.split(",");
-        }
-
-        name = tileLayTag.getAttributeAsString("name");
-
-        extra = tileLayTag.getAttributeAsBoolean("extra", extra);
-        free = tileLayTag.getAttributeAsBoolean("free", free);
-        connected = tileLayTag.getAttributeAsBoolean("connected", connected);
-        discount = tileLayTag.getAttributeAsInteger("discount", discount);
-
-        if (tileId != null) {
-            description = LocalText.getText("LayNamedTileInfo",
-                    tileId,
-                    name != null ? name : "",
-                            locationCodes,
-                            (extra ? LocalText.getText("extra"):LocalText.getText("notExtra")),
-                            (free ? LocalText.getText("noCost") : discount != 0 ? LocalText.getText("discount", discount) : 
-                                LocalText.getText("normalCost")),
-                            (connected ? LocalText.getText("connected") : LocalText.getText("unconnected"))
-            );
-        } else {
-            description = LocalText.getText("LayTileInfo",
-                    locationCodes,
-                    (tileColours != null ? Arrays.toString(tileColours).replaceAll("[\\[\\]]", ""): ""),
-                    (extra ? LocalText.getText("extra"):LocalText.getText("notExtra")),
-                    (free ? LocalText.getText("noCost") : discount != 0 ? LocalText.getText("discount", discount) : 
-                        LocalText.getText("normalCost")),
-                    (connected ? LocalText.getText("connected") : LocalText.getText("unconnected"))
-            );
-        }
-
     }
 
     @Override
-	public void finishConfiguration (RailsRoot root)
-    throws ConfigurationException {
+	public abstract void finishConfiguration(RailsRoot root)
+    throws ConfigurationException;
 
-        TileManager tmgr = root.getTileManager();
-        MapManager mmgr = root.getMapManager();
-        MapHex hex;
-
-        if (tileId != null) {
-            tile = tmgr.getTile(tileId);
-        }
-
-        locations = new ArrayList<MapHex>();
-        for (String hexName : locationCodes.split(",")) {
-            hex = mmgr.getHex(hexName);
-            if (hex == null)
-                throw new ConfigurationException("Location " + hexName
-                        + " does not exist");
-            locations.add(hex);
-        }
-
-    }
-
+    @Override
     public boolean isExecutionable() {
         return true;
     }
@@ -119,7 +56,7 @@ public class SpecialTileLay extends SpecialProperty {
     public boolean isFree() {
         return free;
     }
-    
+
     public int getDiscount() {
         return discount;
     }
@@ -149,20 +86,11 @@ public class SpecialTileLay extends SpecialProperty {
     }
 
     @Override
-	public String toText() {
-        return "SpecialTileLay comp=" + originalCompany.getId()
-        + " hex=" + locationCodes
-        + " colour="+Util.joinWithDelimiter(tileColours, ",")
-        + " extra=" + extra + " cost=" + free + " connected=" + connected;
-    }
+	public abstract String toText();
 
     @Override
-    public String toMenu() {
-        return description;
-    }
+    public abstract String toMenu();
 
     @Override
-    public String getInfo() {
-        return description;
-    }
+    public abstract String getInfo();
 }
