+++ src/main/resources/data/1835/CompanyManager.xml
@@ -52,11 +52,11 @@
 	</Company>
 	<Company name="OBB" longname="Ostbayerische Bahn" type="Private" basePrice="120" revenue="10">
 		<SpecialProperties>
-			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-				<SpecialTileLay location="M15" colour="yellow" extra="yes" free="yes" />
+			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+				<SpecialSingleTileLay location="M15" colour="yellow" extra="yes" free="yes" />
 			</SpecialProperty>
-			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-				<SpecialTileLay location="M17" colour="yellow" extra="yes" free="yes" />
+			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+				<SpecialSingleTileLay location="M17" colour="yellow" extra="yes" free="yes" />
 			</SpecialProperty>
 		</SpecialProperties>
 		<ClosingConditions>
@@ -67,8 +67,8 @@
 	</Company>
 	<Company name="PfB" longname="Pfalzbahnen" type="Private" basePrice="150" revenue="15">
 		<SpecialProperties>
-			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" closesPrivate="no" class="net.sf.rails.game.special.SpecialTileLay">
-				<SpecialTileLay location="L6" colour="green" extra="yes" free="yes" />
+			<SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" closesPrivate="no" class="net.sf.rails.game.special.SpecialSingleTileLay">
+				<SpecialSingleTileLay location="L6" colour="green" extra="yes" free="yes" />
 			</SpecialProperty>
 			<IfOption name="NF_PfB_Westermann" value ="yes">
 				<SpecialProperty condition="ifOwnedByPlayer" when="tokenLayingStep" class="net.sf.rails.game.special.SpecialBaseTokenLay">
