+++ src/main/resources/data/1837/CompanyManager.xml
@@ -70,8 +70,8 @@
         <Company name="KwB" longname="Karawankenbahn" type="Private" basePrice="170" revenue="25">
                 <Blocking hex="J12"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="J12" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="J12" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
                 <Info key="ComesWithPresidency" parm="S2,100"/>
@@ -79,8 +79,8 @@
         <Company name="BrB" longname="Brennerbahn" type="Private" basePrice="140" revenue="15">
                 <Blocking hex="J6"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="J6" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="J6" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
                 <Info key="ComesWithPresidency" parm="S4,100"/>
@@ -88,16 +88,16 @@
         <Company name="WB" longname="Wocheinerbahn" type="Private" basePrice="130" revenue="30">
                 <Blocking hex="K11"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="K11" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="K11" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
         </Company>
         <Company name="AB" longname="Arlbergbahn" type="Private" basePrice="185" revenue="20">
                 <Blocking hex="I5"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="I5" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="I5" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
                 <Info key="ComesWithPresidency" parm="S3,100"/>
@@ -105,8 +105,8 @@
         <Company name="KB" longname="Karstbahn" type="Private" basePrice="120" revenue="10">
                 <Blocking hex="K13"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="K13" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="K13" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
                 <Info key="ComesWithPresidency" parm="S5,100"/>
@@ -114,8 +114,8 @@
         <Company name="SmB" longname="Semmeringbahn" type="Private" basePrice="150" revenue="5">
                 <Blocking hex="H16"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="H16" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="H16" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
                 <Info key="ComesWithPresidency" parm="S1,100"/>
@@ -123,8 +123,8 @@
         <Company name="TB" longname="Tauernbahn" type="Private" basePrice="150" revenue="35">
                 <Blocking hex="J10"/>
                 <SpecialProperties>
-                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-                                <SpecialTileLay location="J10" colour="yellow" extra="no" free="yes"/>
+                        <SpecialProperty condition="ifOwnedByPlayer" when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+                                <SpecialSingleTileLay location="J10" colour="yellow" extra="no" free="yes"/>
                         </SpecialProperty>
                 </SpecialProperties>
         </Company>
