+++ src/main/resources/data/18AL/CompanyManager.xml
@@ -35,8 +35,8 @@
 		longname="Brown &amp; Sons Lumber Co.">
 		<SpecialProperties>
 			<SpecialProperty condition="ifOwnedByCompany"
-				when="tileLayingStep" class="net.sf.rails.game.special.SpecialTileLay">
-				<SpecialTileLay tile="445" name="Lumber Terminal"
+				when="tileLayingStep" class="net.sf.rails.game.special.SpecialSingleTileLay">
+				<SpecialSingleTileLay tile="445" name="Lumber Terminal"
 					location="G2,M2,N5,O4,P5" connected="no" free="yes" extra="yes" />
 			</SpecialProperty>
 		</SpecialProperties>
